requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
ape-alchemy==0.8.0
ape-ens==0.8.0
ape-etherscan==0.8.0
ape-foundry==0.8.0
ape-hardhat==0.8.0
ape-infura==0.8.0
ape-solidity==0.8.2
ape-template==0.8.0
ape-tokens==0.8.0
ape-vyper==0.8.1
apepay==0.3.2
arrow==1.3.0
astroid==3.2.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
base58==1.0.3
binaryornot==0.4.4
bitarray==3.0.0
black==24.4.2
cached-property==2.0.1
cbor2==5.6.4
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chompjs==1.2.4
ckzg==1.0.2
click==8.1.7
cookiecutter==2.1.1
coverage==7.6.1
cryptography==43.0.3
cytoolz==1.0.0
dataclassy==0.11.1
decorator==5.1.1
dill==0.3.8
eip712==0.2.10
eth-account==0.11.3
eth-ape==0.8.19
eth-bloom==3.0.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-pydantic-types==0.1.3
eth-rlp==1.0.1
eth-tester==0.11.0b2
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
ethpm-types==0.6.18
evm-trace==0.2.3
evmchains==0.1.1
exceptiongroup==1.2.2
executing==2.1.0
fief-client==0.20.0
flake8==7.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
iniconfig==2.0.0
ipython==8.29.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jinja2-time==0.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
lazyasd==0.1.4
lru-dict==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
morphys==1.0
msgspec==0.18.6
multidict==6.1.0
mypy==1.10.1
mypy-extensions==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.3
parsimonious==0.10.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cid==0.3.0
py-ecc==7.0.1
py-evm==0.10.1b1
py-geth==5.0.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
py-solc-x==2.0.3
pycodestyle==2.12.0
pycparser==2.22
pycron==3.1.1
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.2.5
pytest==8.3.3
pytest-cov==5.0.0
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rlp==4.0.1
rpds-py==0.21.0
safe-pysha3==1.0.4
semantic-version==2.10.0
silverback==0.6.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
stack-data==0.6.3
taskiq==0.11.7
taskiq-dependencies==1.5.4
termcolor==2.3.0
text-unidecode==1.3
tokenlists==0.1.7
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
toolz==1.0.0
tqdm==4.67.0
traitlets==5.14.3
trie==3.0.1
types-python-dateutil==2.9.0.20240316
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
varint==1.0.2
vvm==0.2.1
vyper==0.3.10
watchdog==3.0.0
wcwidth==0.2.13
web3==6.20.1
websockets==13.1
yarl==1.17.1
yaspin==3.1.0
zipp==3.20.2
